working_requirements.txt
adapter-transformers @ git+https://github.com/Adapter-Hub/adapter-transformers.git@fb98f06bf5f4cca8b5341b27ea2b1ca27db295ff
attrs==21.2.0
black==21.9b0
blis==0.7.4
breezy==3.0.2
catalogue==2.0.6
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==7.1.2
configobj==5.0.6
cymem==2.0.5
Cython==0.29.24
dataclasses==0.6
dbus-python==1.2.16
decorator==5.1.0
deprecation==2.1.0
diskcache==5.2.1
dkpro-cassis==0.5.3
dulwich==0.19.15
fastimport==0.9.8
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
huggingface-hub==0.0.17
idna==2.8
importlib-resources==5.2.2
-e git+https://github.com/inception-project/inception-external-recommender.git@41d894c05053720d2b37510568d11d36433c3cf9#egg=inception_external_recommender
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
lightgbm==3.1.1
lxml==4.6.3
MarkupSafe==2.0.1
more-itertools==8.10.0
murmurhash==1.0.5
mypy-extensions==0.4.3
networkx==2.4
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pathspec==0.9.0
pathy==0.6.0
platformdirs==2.3.0
preshed==3.0.5
protobuf==3.18.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.3.1
PyGObject==3.36.0
pyparsing==2.4.7
python-apt==2.0.0+ubuntu0.20.4.5
python-crfsuite==0.9.7
PyYAML==5.3.1
regex==2021.8.28
requests==2.22.0
requests-unixsocket==0.2.0
rust-fst==0.1.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentence-transformers==0.3.9
sentencepiece==0.1.91
simalign==0.3
six==1.14.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
sortedcontainers==2.4.0
spacy==3.1.2
spacy-legacy==3.0.8
spacy-stanza==1.0.0
srsly==2.4.1
stanza==1.2.3
tabulate==0.8.9
thinc==8.0.10
threadpoolctl==2.2.0
tokenizers==0.10.3
tomli==1.2.1
toposort==1.6
torch==1.7.0
tqdm==4.62.2
transformers==3.5.1
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.25.8
wasabi==0.8.2
Werkzeug==2.0.1
wget==3.2
zipp==3.5.0
